[FLINK-30665][table] Split the test SqlToOperationConverterTest into three different test classes to avoid it exceed the maximum row limit
